+++ client/src/pages/App.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <span>Hello world</span>
+  );
+}
+
+export default App;
